processor/pyproject.toml
[tool.poetry]
name = "processor"
version = "0.1.0"
description = ""
authors = ["Javier Morales de Vera <javiermoralesdevera99@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pendulum = "^3.0.0"
black = "^24.8.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
boto3 = "^1.35.29"
ffmpeg-python = "^0.2.0"
python-dotenv = "^1.0.1"
openai = "^1.51.2"
tiktoken = "^0.8.0"
pandas = "^2.2.3"
numpy = "^2.1.2"
flask = "^3.1.0"
waitress = "^3.0.2"
pytest = "^8.3.3"
yt-dlp = "^2024.12.13"
moviepy = "^2.1.2"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
